back/app/requirements.txt
black
isort
fastapi
httpx
uvicorn
beanie
pytest
pymongo
pydantic
pydantic-settings
cryptography
mongomock-motor
PyJWT
python-multipart
jinja2
pymupdf 
pytesseract 
pillow
pdf2image
pytest
openai
python-magic

#dev 
ipdb
